moved dtos to types package
