--- apps/api/src/printer/printer.dto.ts
@@ -1,17 +1,0 @@
-import { IsNotEmpty, IsNumber } from 'class-validator';
-
-export class PrinterAssignmentDto {
-  @IsNumber()
-  @IsNotEmpty()
-  userId: number;
-
-  @IsNumber()
-  @IsNotEmpty()
-  printerId: number;
-}
-
-export class UserToPrinterDto {
-  id: number;
-  userId: number;
-  printerId: number;
-}
