+++ packages/types/src/dto/printer.ts
@@ -1,0 +1,35 @@
+import { IsNotEmpty, IsNumber, IsString } from 'class-validator';
+
+export class PrinterAssignmentDto {
+  @IsNumber()
+  @IsNotEmpty()
+  userId: number;
+
+  @IsNumber()
+  @IsNotEmpty()
+  printerId: number;
+}
+
+export class PrinterDto {
+  @IsNumber()
+  @IsNotEmpty()
+  id: number;
+
+  @IsString()
+  @IsNotEmpty()
+  name: string;
+}
+
+export class UserToPrinterDto {
+  @IsNumber()
+  @IsNotEmpty()
+  id: number;
+
+  @IsNumber()
+  @IsNotEmpty()
+  userId: number;
+
+  @IsNumber()
+  @IsNotEmpty()
+  printerId: number;
+}
