+++ packages/types/src/index.ts
@@ -1,5 +1,6 @@
 export * from './dto';
 export * from './dto/code';
+export * from './dto/printer';
 export * from './enum';
 
 export * from './dto/registration';
